Introduced rupee bug but fixed other graphics bugs. Perf implications unknown

+++ source/graphics.h
@@ -61,9 +61,9 @@ void SetupScreens();
 void DynamicScreen();
 // Public Sub SuperPrint(SuperWords As String, Font As Integer, X As Single, Y As Single) 'prints text to the screen
 // prints text to the screen
-void SuperPrint(std::string SuperWords, int Font, float X, float Y);
+void SuperPrint(std::string SuperWords, int Font, int X, int Y);
 // for certain utility functions
-void SuperPrint(const char* SuperWords, int Font, float X, float Y);
+void SuperPrint(const char* SuperWords, int Font, int X, int Y);
 // Public Function CheckKey(newStrizzle As String) As String
 //std::string CheckKey(std::string newStrizzle); // USELESS
 // Private Sub SpecialFrames() 'update frames for special things such as coins and kuribo's shoe
